refactor: ストリームインスペクターのイベント収集をコンポーネントから分離
イベント収集がコンポーネントのDOMライフサイクルに結合していたため、
スマホなど1カラム表示環境ではカラムがアンマウントされると収集が停止
していた。Pinia storeに分離し、デッキにカラムが存在する限り
バックグラウンドで収集を継続するようにした。
